Move Shadowy Insight from spells to talents
