+++ src/analysis/retail/priest/shadow/modules/talents/ShadowyInsight.tsx
@@ -28,6 +28,7 @@ class ShadowyInsight extends Analyzer {
 
   constructor(options: Options) {
     super(options);
+    this.active = this.selectedCombatant.hasTalent(TALENTS.SHADOWY_INSIGHT_TALENT);
 
     this.addEventListener(
       Events.applybuff.by(SELECTED_PLAYER).spell(SPELLS.SHADOWY_INSIGHT_BUFF),
